Java FFM instead of JNI
This commit removes the old JNI java bindings and instead uses Java new
Foreign Function and Memory (FFM) API. This API was stabalized as of
Java 21. The Statistics.java class demonstrates how to extend the
AbstractNativeLibrary class and access the imgal's shared object
library. Note that the bindings fetch the imgal library by copying
libimgal.so (for linux) resource into a temporary file on the system
(marked for deletion when the JVM closes). This organization means that
the imgal library is shipped/packaged in the .jar file thats created
when compiled, making distribution of platform specific binaries easier.

This implementation is not complete. In order for Java FFM to work, the
pyO3 Python bindings must be disabled. pyO3 changes the memory layout of
the native library in such a way that is unreadable by
SymbolLookup.libraryLookup(). Sub libraries that depend on a core Rust
library is needed to seperate the Java and Python build requirements.