Type sum function on &ArrayBase<S, D>
This is an example on how to type funcitions on &ArrayBase<S, D> instead of
Array or ArrayView. This has the added advantage of *not* consuming owned arrays
or owned views. Even though an ArrayView doesn't copy the underlying data, there
is still overhead in creating the view itself. Consuming it here is wasteful as
the caller will need to create a new view.

This pattern will be expanded to the rest of the code base.