Add implementation of Shulei's SACA (2D)
This commit adds the SACA algorithm and the supporting
functions. It currently only works with 2D data, but
is easily extendable to 3D. It uses a proper weighted
circle neighborhood and is notably faster. Instead
of kendall's tau-a, it usues kendall's tau-b, removing
the need for a random number generator and making the
algorithm deterministic.

There is a bug in the kendall's tau-b currently
if one of the buffers contains the same value.