Use an epsilon to prevent 0 and smalldenominators
If one of the arrays, say data_a, contains all the same
values then the tie correction becomes very close to
sum weighted pairs. When computing the denominator, these
makes it very small, resulting in a tau value
that 1) exceeds 1.0 and 2) makes no sense.

Because there is no variation in one of the data arrays
the correlation is undefined. Thus a return of 0.0
is appropriate.