Do not create parallel iterator inside a parallel iterator
This was interesting. If you make an parallel iterator
inside another parallel iterator it absolutely hogs
all your threads. Instead of your threads working
on the chunk you want, they're busy playing around
with the overhead of thread work stealing on small arrays.

Be thoughtful about where you do parallel iterators!
Not everything needs to be parallized! Rayon is supposed
to be "sensible" about when to parallize and this
lesson taught me to trust my own senses more.