Remove histogram quality functions
I had a nice chat with a fellow scientist (Jenu Chacko)
about this idea and after some discussion we both
agree it doesn't quite make sense to implement
a filter on the raw data. Also the methodology is
flawed in that the time domain is truncated with
the count_threshold.

Oh well! This is what imgal is for! Trying new things
vetting them and seeing if they're worth keeping
in the library.