Use ofArray and the critical linker option
On heap memory access is not allowed be default for
stability reasons. My guess is that it may interfer
with the JVM's garbage collector. Not too sure. The
critical linker option is intended for "short" functions
that need memory access. I'm unsure what the consequences
are of using the critical option, but I will test.

Ideally, if this works then I don't need to copy
data into of heap memory.